get rid of useless ifd

+++ osu-wine.nix
@@ -18,35 +18,12 @@ with callPackage "${path}/pkgs/applications/emulators/wine/util.nix" { };
 let
   patch = (callPackage ./sources.nix { }).staging;
   build-inputs = pkgNames: extra: (mkBuildInputs wineUnstable.pkgArches pkgNames) ++ extra;
-  collectPatches =
-    d:
-    builtins.concatLists (
-      lib.mapAttrsToList (
-        k: v:
-        (
-          if v == "regular" && lib.hasSuffix ".patch" k then
-            lib.toList
-          else if v == "directory" && !lib.hasPrefix "." k then
-            collectPatches
-          else
-            lib.const [ ]
-        )
-          "${d}/${k}"
-      ) (builtins.readDir d)
-    );
-  # ps0322 breaks linking
-  patchList =
-    map (x: if lib.hasSuffix "/disable-ime-envvar.patch" x then ./disable-ime-envvar.patch else x)
-      (
-        builtins.filter (x: !lib.hasInfix "/ps0322-" x) (
-          collectPatches (fetchFromGitHub {
-            owner = "whrvt";
-            repo = "wine-osu-patches";
-            rev = "15f50183571e8c7068a70454cafd18a14ba56b12";
-            hash = "sha256-EhsRzbarkB5EOnwm/mXybeohhruGGJPTFVJym3PmxZQ=";
-          })
-        )
-      );
+  patches = fetchFromGitHub {
+    owner = "whrvt";
+    repo = "wine-osu-patches";
+    rev = "15f50183571e8c7068a70454cafd18a14ba56b12";
+    hash = "sha256-EhsRzbarkB5EOnwm/mXybeohhruGGJPTFVJym3PmxZQ=";
+  };
   disabled = [ "eventfd_synchronization" ];
 in
 assert lib.versions.majorMinor wineUnstable.version == lib.versions.majorMinor patch.version;
@@ -79,10 +56,17 @@ assert lib.versions.majorMinor wineUnstable.version == lib.versions.majorMinor p
       python3 ./staging/patchinstall.py DESTDIR="$PWD" --all ${
         lib.concatMapStringsSep " " (ps: "-W ${ps}") (patch.disabledPatchsets ++ disabled)
       }
-      for patch in ${builtins.concatStringsSep " " patchList}; do
+      sanityCheck=
+      # ps0322 breaks linking
+      for patch in $(find ${patches} -regex '.*\.patch' | sort | grep -v ps0322); do
+        if [[ "$patch" = */disable-ime-envvar.patch ]]; then
+          patch="${./disable-ime-envvar.patch}"
+          sanityCheck=1
+        fi
         echo "Applying $patch"
         patch -p1 < "$patch"
       done
+      [[ -z "$sanityCheck" ]] && exit 1
       find . -iregex ".*orig" -execdir rm '{'''}' '+' || true
       tools/make_requests
       tools/make_specfiles
